requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
google-generativeai
streamlit
pydantic
python-dotenv
requests
pillow
openai
groq
